rename croning_naist_syllabus to crawling_naist_syllabus
