+++ src/crawling_naist_syllabus/gui.py
