+++ tests/unit/test_fetchdata.py
@@ -1,4 +1,4 @@
-from src.croning_naist_syllabus.fetch import FetchData, LectureNameUrl
+from src.crawling_naist_syllabus.fetch import FetchData, LectureNameUrl
 
 
 def test_fetch_data():
